+++ src/Service/CloudFormation/src/Input/DescribeStacksInput.php
@@ -33,7 +33,7 @@ final class DescribeStacksInput extends Input
     private $stackName;
 
     /**
-     * A string that identifies the next page of stacks that you want to retrieve.
+     * The token for the next set of items to return. (You received this token from a previous call.).
      *
      * @var string|null
      */