+++ src/Service/MediaConvert/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - AWS api-change: This release adds the ability to set resolution for the black video generator.
 - AWS api-change: Adds SlowPalPitchCorrection to audio pitch correction settings. Enables opacity for VideoOverlays. Adds REMUX_ALL option to enable multi-rendition passthrough to VideoSelector for allow listed accounts.
+- AWS api-change: Lowers minimum duration for black video generator. Adds support for embedding and signing C2PA content credentials in DASH and CMAF HLS outputs.
 
 ### Dependency bumped